Fold CompletePatchOutput into VariableDefinition
- Restructure the variable_definitions_files to avoid circular imports
- Simplify test structure
