+++ src/dapla_metadata/variable_definitions/exceptions.py
@@ -186,9 +186,6 @@ def vardef_file_error_handler(method):  # noqa: ANN201, ANN001
         except YAMLError as e:
             msg = f"Invalid yaml. Please fix the formatting in your yaml file.\nOriginal error:\n{e!s}"
             raise VardefFileError(msg) from e
-        except AttributeError as e:
-            msg = f"There is no such attribute. Original error: {e!s}"
-            raise VardefFileError(msg) from e
         except EOFError as e:
             msg = "Unexpected end of file"
             raise VardefFileError(msg) from e
