+++ tests/variable_definitions/test_vardef.py
@@ -29,7 +29,6 @@ from dapla_metadata.variable_definitions.exceptions import STATUS_EXPLANATIONS
 from dapla_metadata.variable_definitions.exceptions import VardefClientError
 from dapla_metadata.variable_definitions.exceptions import VariableNotFoundError
 from dapla_metadata.variable_definitions.vardef import Vardef
-from dapla_metadata.variable_definitions.variable_definition import CompletePatchOutput
 from dapla_metadata.variable_definitions.variable_definition import VariableDefinition
 from tests.utils.constants import VARDEF_EXAMPLE_ACTIVE_GROUP
 from tests.utils.constants import VARDEF_EXAMPLE_DATE
@@ -53,7 +52,7 @@ def test_list_variable_definitions_with_date_of_validity(
     VardefClient.set_config(client_configuration)
     assert isinstance(
         Vardef.list_variable_definitions(date_of_validity=VARDEF_EXAMPLE_DATE)[0],
-        CompletePatchOutput,
+        VariableDefinition,
     )
 
 
@@ -175,7 +174,7 @@ def test_create_draft(
     my_draft = Vardef.create_draft(
         draft=draft,
     )
-    assert isinstance(my_draft, CompletePatchOutput)
+    assert isinstance(my_draft, VariableDefinition)
     assert my_draft.id is not None
     assert my_draft.patch_id == 1
     assert my_draft.variable_status == VariableStatus.DRAFT
@@ -184,7 +183,7 @@ def test_create_draft(
 @pytest.mark.usefixtures("work_folder_complete_patch_output")
 def test_create_draft_from_file():
     my_draft = Vardef.create_draft_from_file()
-    assert isinstance(my_draft, CompletePatchOutput)
+    assert isinstance(my_draft, VariableDefinition)
     assert my_draft.id is not None
     assert my_draft.patch_id == 1
     assert my_draft.variable_status == VariableStatus.DRAFT
@@ -192,7 +191,7 @@ def test_create_draft_from_file():
 
 def test_create_draft_from_file_specify_path(work_folder_complete_patch_output: Path):
     my_draft = Vardef.create_draft_from_file(work_folder_complete_patch_output)
-    assert isinstance(my_draft, CompletePatchOutput)
+    assert isinstance(my_draft, VariableDefinition)
     assert my_draft.id is not None
     assert my_draft.patch_id == 1
     assert my_draft.variable_status == VariableStatus.DRAFT
@@ -207,7 +206,7 @@ def test_migrate_from_vardok(
     my_draft = Vardef.migrate_from_vardok(
         vardok_id="1607",
     )
-    assert isinstance(my_draft, CompletePatchOutput)
+    assert isinstance(my_draft, VariableDefinition)
     assert my_draft.id is not None
     assert my_draft.patch_id == 1
     assert my_draft.variable_status == VariableStatus.DRAFT
