+++ tests/variable_definitions/test_variable_definition_files.py
@@ -10,6 +10,9 @@ import pytz
 import ruamel.yaml
 from pytest_mock import MockerFixture
 
+from dapla_metadata.variable_definitions._generated.vardef_client.models.variable_status import (
+    VariableStatus,
+)
 from dapla_metadata.variable_definitions._utils.config import WORKSPACE_DIR
 from dapla_metadata.variable_definitions._utils.constants import TEMPLATE_HEADER
 from dapla_metadata.variable_definitions._utils.constants import (
@@ -21,16 +24,17 @@ from dapla_metadata.variable_definitions._utils.constants import (
 from dapla_metadata.variable_definitions._utils.constants import (
     TEMPLATE_SECTION_HEADER_STATUS,
 )
-from dapla_metadata.variable_definitions._utils.variable_definition_files import (
+from dapla_metadata.variable_definitions._utils.files import _get_workspace_dir
+from dapla_metadata.variable_definitions._utils.template_files import (
     _find_latest_template_file,
 )
-from dapla_metadata.variable_definitions._utils.variable_definition_files import (
-    _get_workspace_dir,
-)
-from dapla_metadata.variable_definitions._utils.variable_definition_files import (
+from dapla_metadata.variable_definitions._utils.template_files import (
     create_template_yaml,
 )
 from dapla_metadata.variable_definitions.exceptions import VardefFileError
+from dapla_metadata.variable_definitions.vardef import Vardef
+from tests.utils.constants import VARDEF_EXAMPLE_DEFINITION_ID
+from tests.utils.constants import VARDEF_EXAMPLE_SHORT_NAME
 from tests.variable_definitions.conftest import VARIABLE_DEFINITION_DICT
 
 yaml = ruamel.yaml.YAML()
@@ -163,3 +167,48 @@ def test_workspace_is_not_dir(monkeypatch: pytest.MonkeyPatch, set_workspace_not
     monkeypatch.setenv(WORKSPACE_DIR, str(set_workspace_not_dir))
     with pytest.raises(NotADirectoryError):
         _get_workspace_dir()
+
+
+def test_write_existing_variable_to_file():
+    file_name = (
+        Vardef.get_variable_definition_by_id(
+            variable_definition_id=VARDEF_EXAMPLE_DEFINITION_ID,
+        )
+        .to_file()
+        .get_file_path()
+    )
+    with file_name.open(encoding="utf-8") as f:
+        parsed_yaml = yaml.load(f)
+
+    assert parsed_yaml["variable_status"] == VariableStatus.DRAFT
+    assert parsed_yaml["short_name"] == VARDEF_EXAMPLE_SHORT_NAME
+
+
+def test_write_existing_variable_to_file_by_short_name():
+    file_name = (
+        Vardef.get_variable_definition_by_shortname(
+            short_name=VARDEF_EXAMPLE_SHORT_NAME,
+        )
+        .to_file()
+        .get_file_path()
+    )
+    with file_name.open(encoding="utf-8") as f:
+        parsed_yaml = yaml.load(f)
+
+    assert parsed_yaml["variable_status"] == VariableStatus.DRAFT
+    assert parsed_yaml["short_name"] == VARDEF_EXAMPLE_SHORT_NAME
+
+
+def test_shortname_and_id_in_filename():
+    file_name = (
+        Vardef.get_variable_definition_by_id(
+            variable_definition_id=VARDEF_EXAMPLE_DEFINITION_ID,
+        )
+        .to_file()
+        .get_file_path()
+    )
+
+    assert (
+        f"variable_definition_{VARDEF_EXAMPLE_SHORT_NAME}_{VARDEF_EXAMPLE_DEFINITION_ID}_"
+        in str(file_name)
+    )
